organized compiler into modules
